requirements.txt
flask
sumy
PyPDF2
python-dotenv
